fix(#47): TabBar 커스텀 훅 공통 컴포넌트로 변경
